Extract empty section title scanner
