+++ src/database/section.rs
@@ -29,20 +29,6 @@ impl Section {
         format!("#{}", self.human_title().to_kebab_case())
     }
 
-    /// populates the given issues list if this section has an empty title
-    pub fn check_empty_title(&self, path: &Path, issues: &mut Vec<Issue>) {
-        if self.human_title().is_empty() {
-            issues.push(Issue::SectionWithoutHeader {
-                location: Location {
-                    file: path.into(),
-                    line: self.line_number,
-                    start: 0,
-                    end: self.title_text_end(),
-                },
-            });
-        }
-    }
-
     /// populates the given issues list with all sections in this document that don't match the configured sections
     pub fn check_mismatching_title(&self, path: &Path, config: &Config, issues: &mut Vec<Issue>) {
         let section_title = self.human_title();
@@ -225,47 +211,6 @@ mod tests {
         for (give, want) in tests {
             let section = Section::with_title(give);
             assert_eq!(section.anchor(), want);
-        }
-    }
-
-    mod check_empty_title {
-        use crate::{Document, Issue, Location};
-        use indoc::indoc;
-        use std::path::PathBuf;
-
-        #[test]
-        fn empty_title() {
-            let content = indoc! {"
-            # test document
-
-            ###
-            content
-            ###
-            content"};
-            let doc = Document::from_str("test.md", content).unwrap();
-            let mut have = vec![];
-            for section in doc.sections() {
-                section.check_empty_title(&PathBuf::from("test.md"), &mut have);
-            }
-            let want = vec![
-                Issue::SectionWithoutHeader {
-                    location: Location {
-                        file: PathBuf::from("test.md"),
-                        line: 2,
-                        start: 0,
-                        end: 3,
-                    },
-                },
-                Issue::SectionWithoutHeader {
-                    location: Location {
-                        file: PathBuf::from("test.md"),
-                        line: 4,
-                        start: 0,
-                        end: 3,
-                    },
-                },
-            ];
-            pretty::assert_eq!(have, want);
         }
     }
