+++ src/scan/empty_section_content.rs
